Day #1272: Experiments/notes. Explore is not the bottleneck. Bottleneck is definitely training decisions.

+++ side_experiments/attribute_practice/solution/solution_wrapper.h
@@ -33,7 +33,7 @@ public:
 	std::vector<AbstractExperimentState*> experiment_context;
 
 	int num_actions;
-	int curr_impact;
+	double curr_impact;
 
 	AbstractExperimentHistory* experiment_history;
